moved manager classes into manager package
