+++ src/main/java/dansplugins/factionsystem/MedievalFactions.java
@@ -2,6 +2,9 @@ package dansplugins.factionsystem;
 
 import dansplugins.factionsystem.bstats.Metrics;
 import dansplugins.factionsystem.data.PersistentData;
+import dansplugins.factionsystem.managers.ConfigManager;
+import dansplugins.factionsystem.managers.LocaleManager;
+import dansplugins.factionsystem.managers.StorageManager;
 import dansplugins.factionsystem.objects.PlayerActivityRecord;
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
