+++ src/main/java/dansplugins/factionsystem/MedievalFactionsAPI.java
@@ -2,6 +2,7 @@ package dansplugins.factionsystem;
 
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
+import dansplugins.factionsystem.managers.*;
 import dansplugins.factionsystem.utils.ArgumentParser;
 import dansplugins.factionsystem.utils.BlockChecker;
 import dansplugins.factionsystem.utils.UUIDChecker;
