+++ src/main/java/dansplugins/factionsystem/commands/ForceCommand.java
@@ -1,8 +1,8 @@
 package dansplugins.factionsystem.commands;
 
-import dansplugins.factionsystem.LocaleManager;
+import dansplugins.factionsystem.managers.LocaleManager;
 import dansplugins.factionsystem.MedievalFactions;
-import dansplugins.factionsystem.StorageManager;
+import dansplugins.factionsystem.managers.StorageManager;
 import dansplugins.factionsystem.commands.abs.SubCommand;
 import dansplugins.factionsystem.data.PersistentData;
 import dansplugins.factionsystem.objects.Faction;
