+++ src/main/java/dansplugins/factionsystem/commands/abs/SubCommand.java
@@ -3,6 +3,9 @@ package dansplugins.factionsystem.commands.abs;
 import dansplugins.factionsystem.*;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
+import dansplugins.factionsystem.managers.ChunkManager;
+import dansplugins.factionsystem.managers.ConfigManager;
+import dansplugins.factionsystem.managers.LocaleManager;
 import dansplugins.factionsystem.objects.Faction;
 import dansplugins.factionsystem.utils.ArgumentParser;
 import org.bukkit.Bukkit;
