+++ src/main/java/dansplugins/factionsystem/eventhandlers/InteractionHandler.java
@@ -3,6 +3,10 @@ package dansplugins.factionsystem.eventhandlers;
 import dansplugins.factionsystem.*;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
+import dansplugins.factionsystem.managers.ChunkManager;
+import dansplugins.factionsystem.managers.GateManager;
+import dansplugins.factionsystem.managers.LocaleManager;
+import dansplugins.factionsystem.managers.LockManager;
 import dansplugins.factionsystem.objects.ClaimedChunk;
 import dansplugins.factionsystem.objects.Faction;
 import dansplugins.factionsystem.objects.Gate;
