+++ src/main/java/dansplugins/factionsystem/eventhandlers/MoveHandler.java
@@ -1,8 +1,8 @@
 package dansplugins.factionsystem.eventhandlers;
 
-import dansplugins.factionsystem.ChunkManager;
+import dansplugins.factionsystem.managers.ChunkManager;
 import dansplugins.factionsystem.DynmapIntegrator;
-import dansplugins.factionsystem.LocaleManager;
+import dansplugins.factionsystem.managers.LocaleManager;
 import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.PersistentData;
 import dansplugins.factionsystem.objects.Faction;
