+++ src/main/java/dansplugins/factionsystem/managers/ChunkManager.java
@@ -1,5 +1,7 @@
-package dansplugins.factionsystem;
+package dansplugins.factionsystem.managers;
 
+import dansplugins.factionsystem.MedievalFactions;
+import dansplugins.factionsystem.Messenger;
 import dansplugins.factionsystem.data.EphemeralData;
 import dansplugins.factionsystem.data.PersistentData;
 import dansplugins.factionsystem.events.FactionClaimEvent;
