+++ src/main/java/dansplugins/factionsystem/managers/ConfigManager.java
@@ -1,8 +1,8 @@
-package dansplugins.factionsystem;
+package dansplugins.factionsystem.managers;
 
+import dansplugins.factionsystem.MedievalFactions;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
 
 public class ConfigManager {
