+++ src/main/java/dansplugins/factionsystem/managers/StorageManager.java
@@ -1,9 +1,10 @@
-package dansplugins.factionsystem;
+package dansplugins.factionsystem.managers;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
+import dansplugins.factionsystem.MedievalFactions;
 import dansplugins.factionsystem.data.PersistentData;
 import dansplugins.factionsystem.objects.*;
